compresseed the boot module into one file
In the future (i.e. TODO), when an allocator is present, we can
use `Arc<Mutex<TYPE>>` for proper multi-threading safe operation.
